better way to manage requests
The UI now keeps tracks of all requests that are pending/sent out.
for every request going out, We check the overlap of the new request
range with the range of the cache when its first initialized. If the overlap
is less than 25%, the cache is cleared and new data is requested. this also
assumes this range might be in the next zoom level.